Split JS Classes
