+++ Crome_JS_Server_V0.1/mainPage.html
@@ -21,5 +21,5 @@
 			
 		</div>
 	</body>
-	<script src="main.js"></script>
+	<script type="module" src="main.js"></script>
 </html>
